nexzy-backend/requirements.txt
# Nexzy Backend Dependencies
# Python FastAPI backend for OSINT credential leak detection

# ==============================================================================
# WEB FRAMEWORK
# ==============================================================================
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# ==============================================================================
# DATABASE & AUTHENTICATION
# ==============================================================================
supabase>=2.0.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# ==============================================================================
# WEB SCRAPING
# ==============================================================================
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
html5lib>=1.1

# ==============================================================================
# OPTIONAL: Advanced Scraping (uncomment if needed)
# ==============================================================================
# selenium==4.16.0
# webdriver-manager==4.0.1
# PySocks==1.7.1  # For Tor support

# ==============================================================================
# UTILITIES
# ==============================================================================
python-multipart>=0.0.6
aiofiles>=23.0.0

# ==============================================================================
# TESTING (optional)
# ==============================================================================
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# ==============================================================================
# DEVELOPMENT (optional)
# ==============================================================================
# black==23.12.1
# flake8==7.0.0
# mypy==1.8.0
